Design class logging as base class. Impove  testing.
The logging functionality for a class, i.e. a logger as attribute with default handlers is now provided as
a class. This class can be inherited to any class that
should have such a logger.

Tests have been adjusted. I introduced mocking
behaviour instead of actually creating log
directories and files. The focus is more on whether
these functions are called rather than that they are
executed.